refactor(snapping): Move back to utils
