+++ lib/components/modification/edit-alignment.js
@@ -14,7 +14,7 @@ import colors from '../../constants/colors'
 import {Button} from '../buttons'
 import {Checkbox, Number} from '../input'
 import type {MapState, GTFSStop} from '../../types'
-import getExistingStopsAlongPattern from '../../selectors/existing-stops-along-pattern'
+import getExistingStopsAlongPattern from '../../utils/get-existing-stops-along-pattern'
 
 type Props = {
   disabled: boolean,
@@ -225,7 +225,13 @@ export default class EditAlignment extends PureComponent {
             onChange={this._onStopSpacingChange}
             units='meters'
           />}
-
+        {isEditing &&
+          <Button block
+            onClick={this._autoGen}
+            style='warning'>
+            <Icon type='circle' /> message('transitEditor.snap')
+          </Button>
+        }
         {modification.type !== REROUTE &&
           <Checkbox
             checked={modification.bidirectional}
