+++ lib/utils/get-existing-stops-along-pattern.js
